Optimized client bandwidth usage

+++ src/cubetech/net/CLC.java
@@ -11,10 +11,10 @@ package cubetech.net;
  */
 public class CLC {
     // Opcodes
-    public static final int OPS_BAD = 0;
-    public static final int OPS_NOP = 1;
-    public static final int OPS_MOVE = 2;
-    public static final int OPS_MOVENODELTA = 3;
-    public static final int OPS_CLIENTCOMMAND = 4;
-    public static final int OPS_EOF = 5;
+    public static final byte OPS_BAD = 0;
+    public static final byte OPS_NOP = 1;
+    public static final byte OPS_MOVE = 2;
+    public static final byte OPS_MOVENODELTA = 3;
+    public static final byte OPS_CLIENTCOMMAND = 4;
+    public static final byte OPS_EOF = 5;
 }
